==> scripts/00_SFA2_dnaconc_spikecalc.Rmd <==
---
title: "SFA2 - DNA concentrations"
author: "Cassandra Wattenburger"
date: "4/7/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(warning = FALSE)
```

```{r}
rm(list=ls())

library(tidyverse)

sessionInfo()
```

# Import data

```{r}
# Create list of files in directory
files <- list.files(path="../data_picogreen/SFA2", full.names=TRUE, recursive=FALSE)

# Import DNA conc. data
dna <- tibble()
for (i in files) {
  if (str_detect(i, c("tidy", ".txt")) & !str_detect(i, "redo")) { # select files tidied using picogreen_tidy.py
    dat <- read_tsv(i) %>%
      select(Sample, AdjConc) %>%
      mutate(Sample = as.double(Sample))
    dna <- bind_rows(dna, dat)
  } else {
    next
  }
}

# Add metadata
meta <- read_tsv("../data_amplicon/SFA2_full/SFA2_full_metadata.tsv")
dna_meta <- full_join(dna, meta) %>%
  arrange(Sample)
```

# Visualize DNA concentrations

```{r}
dna_meta %>%
  mutate(Innoculant = as_factor(Innoculant)) %>%
  ggplot(aes(x=Day, y=AdjConc)) +
  geom_point(aes(color=Innoculant)) +
  #facet_wrap(~Innoculant) +
  labs(y="DNA conc. (ng/uL)") +
  theme_test() 

# Graph each innoculant separately with sample labels
for (x in as.character(unique(dna_meta$Innoculant))) {
    p <- dna_meta %>%
    filter(Innoculant == x) %>%
    ggplot(aes(x=Day, y=AdjConc)) +
    geom_point(alpha=0.5) +
    geom_text(aes(label=Sample), size=3, hjust=0, vjust=0) +
    labs(y="DNA conc. (ng/uL)", title=paste("Innoculant", x)) +
    theme_test()
    print(p)
}

hist(dna_meta$AdjConc, breaks=15)

```

# Incoprorate redos

```{r}
# Import DNA conc. data
for (i in files) {
  if (str_detect(i, c("tidy", "redos", ".txt"))) { # select files tidied using picogreen_tidy.py
    dna.redo <- read_tsv(i) %>%
      select(Sample, AdjConc) %>%
      mutate(Sample = as.double(Sample)) %>%
      add_column(Status = rep("redo", nrow(.)))
  } else {
    next
  }
}

# Add metadata
redo.meta <- inner_join(dna.redo, meta) %>%
  arrange(Sample)

# First pass DNA extractions
dna_meta <- add_column(dna_meta, Status = rep("original", nrow(dna_meta)))

# Add redos
dna_meta <- bind_rows(dna_meta, redo.meta)
```


```{r}
redos <- c(163, 164, 165, 166, 218, 219, 220, 235, 236, 237, 252, 253, 269, 271, 275, 276, 277, 280, 281, 282, 287, 288, 310, 312, 325, 341)

dna_meta %>%
  filter(Sample %in% redos) %>%
  mutate(Sample = as.factor(Sample)) %>%
  ggplot(aes(x=Sample, y=AdjConc)) +
  geom_point(aes(color=Status)) +
  labs() +
  theme_test()

dna_meta %>%
  filter(Sample %in% redos) %>%
  mutate(Sample = as.factor(Sample)) %>%
  ggplot(aes(y=AdjConc)) +
  geom_boxplot(aes(color=Status)) +
  labs() +
  theme_test()
```

Redo DNA extraction seems to have improved yield overall.

# Visualize DNA concentrations with redos subbed in

```{r}
# Graph each innoculant over time
for (x in as.character(unique(dna_meta$Innoculant))) {
    p <- dna_meta %>%
    filter(!(Status == "original" & Sample %in% redos)) %>%
    filter(Innoculant == x) %>%
    ggplot(aes(x=Day, y=AdjConc)) +
    geom_point(alpha=0.5) +
    geom_text(aes(label=Sample), size=3, hjust=0, vjust=0) +
    labs(y="DNA conc. (ng/uL)", title=paste("Innoculant", x)) +
    theme_test()
    print(p)
}
  
```

Compare DNA over time of high and low treatments:

```{r}
dna_meta %>% 
  filter(DOC %in% c("high", "low")) %>% 
  ggplot(aes(x=Day, y=AdjConc)) +
  geom_jitter(aes(color=DOC)) +
  theme_test()

dna_meta %>% 
  filter(DOC %in% c("high", "low")) %>% 
  ggplot(aes(x=DOC, y=AdjConc)) +
  geom_boxplot(aes(color=DOC)) +
  facet_wrap(~Day, scales = "free") +
  theme_test()
```

# Identify a threshold for spike-in levels

Average DNA concentrations by inncoulants and day.

```{r}
# Replace redos
dna_meta_sub <- filter(dna_meta, !(Status == "original" & Sample %in% redos))

# Average
dna_avg <- dna_meta_sub %>%  
  group_by(Innoculant, Day, DOC_pheno) %>%
  summarize(dna_avg = mean(AdjConc), dna_sd = sd(AdjConc))
  
# Visualize
dna_avg %>%
  mutate(Innoculant = as.factor(Innoculant)) %>% 
  filter(!Day == 30) %>%
  ggplot(aes(x=Day, y=dna_avg)) +
  geom_point(aes(color=Innoculant)) +
  geom_line(aes(color=Innoculant)) +
  geom_hline(yintercept=1, linetype=2) +
  labs(y="Mean DNA conc. (ng/uL)") +
  theme_test()

# log transform
dna_avg %>%
  mutate(Innoculant = as.factor(Innoculant)) %>% 
  filter(!Day == 30) %>%
  ggplot(aes(x=Day, y=log(dna_avg))) +
  geom_point(aes(color=Innoculant)) +
  geom_line(aes(color=Innoculant)) +
  facet_wrap(~Innoculant) +
  geom_vline(xintercept = 10, linetype = 2) +
  labs(y="Log mean DNA conc. (ng/uL)") +
  theme_test()

# separate by DOC phenotype
dna_avg %>%
  mutate(Innoculant = as.factor(Innoculant)) %>% 
  filter(!Day == 30) %>%
  ggplot(aes(x=Day, y=dna_avg)) +
  geom_point(aes(color=Innoculant)) +
  geom_line(aes(color=Innoculant)) +
  facet_wrap(~DOC_pheno) +
  geom_hline(yintercept=1, linetype=2) +
  labs(title="Separared by DOC phenotype", y="Mean DNA conc. (ng/uL)") +
  theme_test()

dna_avg %>%
  mutate(Innoculant = as.factor(Innoculant)) %>% 
  filter(!Day == 30) %>%
  ggplot(aes(x=Day, y=log(dna_avg))) +
  geom_point(aes(color=Innoculant)) +
  geom_line(aes(color=Innoculant)) +
  facet_wrap(~DOC_pheno) +
  labs(title="Separared by DOC phenotype", y="Log mean DNA conc. (ng/uL)") +
  theme_test()
```

Range of values above and below threshold of 1, averaged.

All data:
```{r}
min(dna_avg$dna_avg)
max(dna_avg$dna_avg)
```

Below 1 ng/uL:
```{r}
# min
dna_avg %>%
  filter(dna_avg < 1) %>%
  group_by() %>%
  summarize(min(dna_avg))

# max
dna_avg %>%
  filter(dna_avg < 1) %>%
  group_by() %>%
  summarize(max(dna_avg))
```

Above 1 ng/uL:
```{r}
# min
dna_avg %>%
  filter(dna_avg > 1) %>%
  group_by() %>%
  summarize(min(dna_avg))

# max
dna_avg %>%
  filter(dna_avg > 1) %>%
  group_by() %>%
  summarize(max(dna_avg))
```

Range of values above and below threshold of 1, raw.

All data:
```{r}
min(dna_meta_sub$AdjConc)
max(dna_meta_sub$AdjConc)
```

Below 1 ng/uL:
```{r}
# min
dna_meta_sub %>%
  filter(AdjConc < 1) %>%
  group_by() %>%
  summarize(min(AdjConc))

# max
dna_meta_sub %>%
  filter(AdjConc < 1) %>%
  group_by() %>%
  summarize(max(AdjConc))
```

Above 1 ng/uL:
```{r}
# min
dna_meta_sub %>%
  filter(AdjConc > 1) %>%
  group_by() %>%
  summarize(min(AdjConc))

# max
dna_meta_sub %>%
  filter(AdjConc > 1) %>%
  group_by() %>%
  summarize(max(AdjConc))
```

If 1 ng/uL is the threshold, each spike-group will vary in DNA concentration by a factor of 10.

Divide data by threshold and calculate spike-in for each group
------------------------

Separate using the averaged dna concentrations so that replicates are kept within the same spike-group.

```{r}
# label groups based on averaged dna concentration
dna_avg_groups <- mutate(dna_avg, spike_group = if_else(dna_avg < 1, "low", "high"))

# visualize
dna_avg_groups %>%
  ggplot(aes(x=spike_group, y=dna_avg)) +
  geom_boxplot() +
  theme_test()

# Apply labels to unaveraged dataset
dna_meta_group <- dna_meta_sub %>%
  full_join(dna_avg_groups) %>%
  select(-dna_avg)

# visualize
dna_meta_group %>%
  ggplot(aes(x=spike_group, y=AdjConc)) +
  geom_boxplot() +
  theme_test()
```

```{r}
dna_meta_group %>%
  group_by(spike_group) %>%
  filter(Day > 0) %>%
  summarize(dna_avg = mean(AdjConc), dna_median = median(AdjConc), dna_min = min(AdjConc), dna_max = max(AdjConc))
```

```{r}
dna_meta_group %>%
  ggplot(aes(x=Day, y=AdjConc)) +
  geom_point(aes(color=spike_group)) +
  facet_wrap(~Innoculant) +
  theme_test()

for (x in as.character(unique(dna_meta_group$Innoculant))) {
    p <- dna_meta_group %>%
    filter(!(Status == "original" & Sample %in% redos)) %>%
    filter(Innoculant == x) %>%
    ggplot(aes(x=Day, y=AdjConc, color=spike_group)) +
    geom_point(alpha=0.5) +
    geom_hline(yintercept = 1) +
    geom_text(aes(label=Sample), size=3, hjust=0, vjust=0) +
    labs(y="DNA conc. (ng/uL)", title=paste("Innoculant", x)) +
    theme_test()
    print(p)
}

# label borderline samples
dna_meta_group <- mutate(dna_meta_group, borders = case_when(spike_group=="low" & AdjConc <= 1 ~ "low",
                                                                  spike_group=="high" & AdjConc >= 1 ~ "high",
                                                                  spike_group=="low" & AdjConc >= 1 ~ "both",
                                                                  spike_group=="high" & AdjConc <= 1 ~ "both"))

# Number of samples in each spike group
dna_meta_group %>%
  group_by(borders) %>%
  summarize(count(.))
```

```{r, eval=FALSE}
write_csv(dna_meta_group, "../data_picogreen/SFA2_dnaspikes.csv")
```


Sample dilutions
----------------

If the sample is less than 1 ng/uL, no dilution. If the sample is > 1 ng/uL, dilute to 1 ng/uL.

This is to prevent bias during PCR due to unequal template addition. Unfortunately, I can't add enough DNA to bring the low concentration samples up to 2 ng in the reaction because I am low on volume. This is a compromise.

```{r}
# Calculate dilutions for each sample
dna_meta_sub <- dna_meta_sub %>%
  mutate(conc_goal = if_else(AdjConc <= 1, AdjConc, 1),
         template_uL = round((conc_goal*20)/AdjConc, 0), # C1V1=C2V2, rounded to nearest whole uL
         mgH2O_uL = 20-template_uL) %>%
  arrange(Sample)
```

```{r, eval=FALSE}
dna_meta_sub %>%
  select(Sample, template_uL, mgH2O_uL) %>%
  write_csv(., file="../data_intermediate/SFA2_dnadilutions.csv")
```


New spike-in calculations
-------------------------

```{r}
# Find high DNA outliers
# My previous library had a problem with having way too much DNA, so want to remove anything that will raise DNA average too much
dna_meta_sub %>%
  ggplot(aes(x=Sample, y=AdjConc)) +
  geom_point() +
  geom_text(aes(label=Sample), vjust="outward") +
  theme_test()

outliers <- c(338, 290, 327, 363, 272, 303)

# Sample DNA mean excluding day 0 (mostly plant DNA) and high outliers
sample_dna_mean <- dna_meta_sub %>% 
  filter(Day != 0, !(Sample %in% outliers)) %>%
  group_by() %>%
  summarize(mean = mean(AdjConc, na.rm=TRUE))

# Day 0 mean
day0_dna_mean <- dna_meta_sub %>%
  filter(Day==0) %>%
  group_by() %>%
  summarize(mean = mean(AdjConc, na.rm=TRUE))

# Subtract day 0 mean from sample average
# This number will be used to calculate new spike-in
actual_dna_mean <- sample_dna_mean - day0_mean

# Calculate total DNA in each sample based on the average
total_dna_mean <- sample_dna_mean*15*1000 # 15 uL, convert from ng to pg

# Calculate amount of spike to reach 0.0001% of total DNA
total_spike_pg <- total_dna_mean*0.000001
total_spike_pg

# Spike concentration to make (plan to add 5 uL to each sample)
spike_conc <- total_spike_pg/5

# Total volume of spike needed for all samples (8 uL to account for liquid loss during pipetting)
vol_total <- 8*384

# Volume needed to dilute spike from 0.1 pg/uL
spike_vol <- (vol_total*spike_conc)/0.1
spike_vol
water_vol <- vol_total - spike_vol
water_vol

# Prepare spike-in separately for each replicate of experiment
spike_vol/3
water_vol/3
```
