==> scripts/01_SFA_internal_std.Rmd <==
---
title: "SFA2 - Internal standard check"
author: "Cassandra Wattenburger"
date: "9/15/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(results = "show")
knitr::opts_chunk$set(message = FALSE)
```

```{r}
# Clear working directory, load in packages, generate package info
rm(list=ls())

library("tidyverse")

sessionInfo()
```

### Import count and taxonomy data

```{r}
# Rep 1
count_rep1 <- read_tsv(file="../data_amplicon/SFA2_full/rep1/final/SFA2_rep1.counts-final.tsv")
tax_rep1 <- read_tsv(file="../data_amplicon/SFA2_full/rep1/final/SFA2_rep1.taxonomy-final.tsv")

# Rep 2
count_rep2 <- read_tsv(file="../data_amplicon/SFA2_full/rep2/final/SFA2_rep2.counts-final.tsv")
tax_rep2 <- read_tsv(file="../data_amplicon/SFA2_full/rep2/final/SFA2_rep2.taxonomy-final.tsv")

# Rep 3
count_rep3 <- read_tsv(file="../data_amplicon/SFA2_full/rep3/final/SFA2_rep3.counts-final.tsv")
tax_rep3 <- read_tsv(file="../data_amplicon/SFA2_full/rep3/final/SFA2_rep3.taxonomy-final.tsv")

# Redos
count_redos <- read_tsv(file="../data_amplicon/SFA2_full/redos/final/SFA2_redos.counts-final.tsv")
tax_redos <- read_tsv(file="../data_amplicon/SFA2_full/redos/final/SFA2_redos.taxonomy-final.tsv")

# Metadata
meta <- read_tsv(file = "../data_amplicon/SFA2_full/SFA2_metadata_corrected.tsv")
```

### Isolate Aquifex ASVs

Create function that isolates Aquifex ASVs from count data:

```{r}
# Define function to isolate Aquifex ASVs
aquifex_count <- function(tax, count) { 
  asvs <- tax %>% 
    filter(Genus == "Aquifex") %>% 
    select(ASV)
  aqui_count <- count %>% 
           filter(ASV %in% asvs$ASV)
}
```

Generate Aquifex count tables:

```{r}
# Isolate Aquifex spike-in ASVs from each dataset
aqui_count_rep1 <- aquifex_count(tax_rep1, count_rep1)
aqui_count_rep2 <- aquifex_count(tax_rep2, count_rep2)
aqui_count_rep3 <- aquifex_count(tax_rep3, count_rep3)
aqui_count_redos <- aquifex_count(tax_redos, count_redos)

# Merge and reformat
meta_reps <- filter(meta, Library %in% c(1, 2, 3))
aqui_count_reps <- inner_join(aqui_count_rep1, aqui_count_rep2) %>% # combine replicates
  inner_join(aqui_count_rep3) %>%
  pivot_longer(cols = -ASV, names_to = "SampleID", values_to = "count") %>% # long format
  mutate(SampleID = as.numeric(SampleID)) %>%
  inner_join(meta_reps, by="SampleID") # add metadata # filter attempt original only

# Redos (can't add using inner_join because repeated sample names)
# problem, counts the same for original and redo samples, replaced original with redo
aqui_asvs <- unique(aqui_count_reps$ASV)
meta_redos <- filter(meta, Library == 4)
aqui_count <- filter(aqui_count_redos, ASV %in% aqui_asvs) %>% # remove Aquifex ASVs not found in other replicates
  pivot_longer(cols = -ASV, names_to = "SampleID", values_to = "count") %>% # long format
  mutate(SampleID = as.numeric(SampleID)) %>%
  inner_join(meta_redos, by="SampleID") %>% # add metadata
  bind_rows(aqui_count_reps) # add to other data

```

### Calculate proportion of Aquifex internal standard in each sample.

Create function to calculate total reads per sample:

```{r}
# Function to calculate total reads per sample
total_reads <- function(count, meta) {
  count %>% 
    gather(SampleID, value, -ASV) %>% # transpose
    mutate(SampleID = as.numeric(SampleID)) %>% 
    spread(ASV, value) %>%
    mutate(total = rowSums(.[-1])) %>% # calculate row sums
    select(SampleID, total) %>% 
    inner_join(meta, by="SampleID")
}
```

Add sample read totals to Aquifex count data:

```{r}
# Calculate total reads per sample
total_rep1 <- total_reads(count_rep1, meta_reps)
total_rep2 <- total_reads(count_rep2, meta_reps)
total_rep3 <- total_reads(count_rep3, meta_reps)
total_redos <- total_reads(count_redos, meta_redos)

# Merge datasets
total_reads <- bind_rows(total_rep1, total_rep2) %>%
  bind_rows(total_rep3) %>%
  bind_rows(total_redos) %>% 
  mutate(SampleID = as.numeric(SampleID))

aqui_total_reads <- inner_join(aqui_count, total_reads) %>%
  select(SampleID, Attempt, count, total, everything(), -c(ASV, BarcodeSequence, Primer_num))
```

Calculate proportion of Aquifex reads per sample:

```{r}
aqui_prop <- mutate(aqui_total_reads, prop = count/total)
```

Summary statistics:

Minimum (not including 0s), maximum, average.

```{r}
aqui_prop %>% filter(prop > 0) %>% arrange(prop) %>% slice(1) %>% .$prop*100
max(aqui_prop$prop)*100
mean(aqui_prop$prop)*100
sd(aqui_prop$prop)*100
```

How many samples have no spike-in? Which ones?

```{r}
aqui_prop %>% 
  filter(prop == 0)
```

Visualize:

```{r}
aqui_prop %>%
  filter(Type != "pcr_control") %>%
  mutate(Replicate = as.factor(Replicate)) %>%
  ggplot(aes(x=SampleID, y=prop, color=Replicate)) +
  geom_point() +
  labs(y="Proportion Aquifex ASV") +
  theme_test()

aqui_prop %>%
  filter(Type != "pcr_control") %>%
  mutate(Replicate = as.factor(Replicate)) %>%
  ggplot(aes(x=total, y=prop, color=Replicate)) +
  geom_point() +
  labs(y="Proportion Aquifex ASV") +
  theme_test()
```
