==> scripts/05_paprica_tidy.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(results = "show")
knitr::opts_chunk$set(message = FALSE)
```

```{r}
# Clear working directory, load in packages, generate package info
rm(list=ls())

library("tidyverse")

sessionInfo()
```

# Import data

Files produced using paprica v0.6. See /home/cassi/SFA/paprica_analysis directory.

```{r}
# PAPRICA outputs
# Bacteria
pap_bac <- read_csv("../paprica_analysis/output_bacteria/subset_rep-seqs.bacteria.edge_data.csv")
place_bac <- read_csv("../paprica_analysis/output_bacteria/subset_rep-seqs.bacteria.combined_16S.bacteria.tax.placements.csv")

# Archaea
pap_arc <- read_csv("../paprica_analysis/output_archaea/subset_rep-seqs.archaea.edge_data.csv")
place_arc <- read_csv("../paprica_analysis/output_archaea/subset_rep-seqs.archaea.combined_16S.archaea.tax.placements.csv")

# Growth and death estimates
growth <- readRDS("../data_intermediate/SFA2_growth_estimates.rds")
death <- readRDS("../data_intermediate/SFA2_death_estimates.rds")

# Taxonomy
tax <- readRDS("../data_intermediate/SFA2_tax_normalized.rds")
```

# Tidy and combine

```{r}
# What's going on here:
# The global_edge_num col in the "...placements.csv" file matches to part of the string in the Pquery col in the "...edge_data.csv" file
# Using these to merge files

# Bacteria
place_bac <- place_bac %>% mutate(ASV = gsub("[0-9]+\\|(*)+", "\1", Pquery),
         ASV = gsub("\001", "", ASV)) %>% 
  select(ASV, global_edge_num)

pap_bac <- pap_bac %>% select(global_edge_num='...1', n16S, genome_size)

paprica_bac <- full_join(place_bac, pap_bac)

# Archaea
place_arc <- place_arc %>% mutate(ASV = gsub("[0-9]+\\|(*)+", "\1", Pquery),
         ASV = gsub("\001", "", ASV)) %>% 
  select(ASV, global_edge_num)

pap_arc <- pap_arc %>% select(global_edge_num='...1', n16S, genome_size)

paprica_arc <- full_join(place_arc, pap_arc)

# Together
pap_all <- bind_rows(paprica_bac, paprica_arc) %>% 
  select(-global_edge_num) %>% 
  inner_join(tax)
```

In the bacterial data, there are several taxa missing in the "...edge_data.csv" file. I'm not sure why, probably an obscure glitch in PAPRICA's pipeline. Submitted an issue [here](https://github.com/bowmanjeffs/paprica/issues/91). I'm going to average the n16S and genome_size across taxa from the same phyla as an approximation.

```{r}
# Missing data
pap_all %>% 
  filter(Phylum == "Acidobacteriota" | Phylum == "Chloroflexi")

# Average across phyla
missing_replace <- pap_all %>% 
  group_by(Phylum) %>% 
  summarize(n16s = mean(n16S, na.rm=TRUE),
            genome_size = mean(genome_size, na.rm=TRUE)) %>% 
  filter(Phylum=="Acidobacteriota" | Phylum=="Chloroflexi")
missing_replace

acido_n16S <- missing_replace[1,2]
acido_genome <- missing_replace[1,3]
chloro_n16S <- missing_replace[2,2]
chloro_genome <- missing_replace[2,3]


# BOOKMARK


# Add back to data
pap_all <- pap_all %>% 
  mutate(n16S = if_else(Phylum=="Acidobacteriota" & is.na(n16S), acido_n16S, if_else(Phylum=="Chloroflexi" & is.na(n16S), chloro_n16S, n16S)),
         genome_size = if_else(Phylum=="Acidobacteriota" & is.na(genome_size), acido_genome, if_else(Phylum=="Chloroflexi" & is.na(genome_size), chloro_genome, genome_size)))
```


# Add PAPRICA output to growth and death estimates

```{r}
growth_pap <- left_join(growth, pap_all, by="ASV")
death_pap <- left_join(death, pap_all, by="ASV")
```

# Save

```{r, eval=FALSE}
saveRDS(growth_pap, "../data_intermediate/SFA2_growth_estimates.rds")
saveRDS(death_pap, "../data_intermediate/SFA2_death_estimates.rds")
```
