==> scripts/05_SFA2_paprica.Rmd <==
---
title: "Untitled"
author: "Cassandra Wattenburger"
date: "3/25/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(results = "show")
knitr::opts_chunk$set(message = FALSE)
```

```{r}
# Clear working directory, load in packages, generate package info
rm(list=ls())

library("tidyverse")

sessionInfo()
```

# Import data

Files produced by paprica v0.6.

```{r}
# PAPRICA outputs
# Bacteria
pap_bac <- read_csv("../paprica_analysis/asvs_all/SFA2_full.rep-seqs-merged.bacteria.edge_data.csv")
place_bac <- read_csv("../paprica_analysis/asvs_all/SFA2_full.rep-seqs-merged.bacteria.combined_16S.bacteria.tax.placements.csv")

# Archaea
pap_arc <- read_csv("../paprica_analysis/asvs_all/SFA2_full.rep-seqs-merged.archaea.edge_data.csv")
place_arc <- read_csv("../paprica_analysis/asvs_all/SFA2_full.rep-seqs-merged.archaea.combined_16S.archaea.tax.placements.csv")

# Growth and death estimates
growth <- readRDS("../data_intermediate/SFA2_growth_estimates2.rds")
death <- readRDS("../data_intermediate/SFA2_death_estimates2.rds")

# Taxonomy
tax <- readRDS("../data_intermediate/SFA2_taxonomy.rds")

# Pathways
path_bac <- read_csv("../paprica_analysis/asvs_all/SFA2_full.rep-seqs-merged.archaea.pathways.csv")
path_arc <- read_csv("../paprica_analysis/asvs_all/SFA2_full.rep-seqs-merged.bacteria.pathways.csv")
```

# PAPRICA results

```{r}
# What's going on here:
# The global_edge_num col in the "...placements.csv" file matches to part of the string in the Pquery col in the "...edge_data.csv" file
# Using these to merge files

# Bacteria
place_bac <- place_bac %>% mutate(ASV = gsub("[0-9]+\\|(*)+", "\1", Pquery),
         ASV = gsub("\001", "", ASV))

pap_bac <- pap_bac %>% select(global_edge_num='...1', n16S, genome_size, phi, npaths_actual, nec_actual, confidence, branch_length)

paprica_bac <- full_join(place_bac, pap_bac)

# Archaea
place_arc <- place_arc %>% mutate(ASV = gsub("[0-9]+\\|(*)+", "\1", Pquery),
         ASV = gsub("\001", "", ASV))

pap_arc <- pap_arc %>% 
  select(global_edge_num='...1', n16S, genome_size, phi, npaths_actual, nec_actual, confidence, branch_length)

paprica_arc <- full_join(place_arc, pap_arc)

# Merge
pap_all <- bind_rows(paprica_bac, paprica_arc) %>% 
  inner_join(tax) %>% 
  select(ASV, global_edge_num, n16S, genome_size, phi, npaths_actual, nec_actual, confidence, branch_length, Domain:Species)
```

### Overall confidence of predictions

```{r}
mean(pap_all$confidence, na.rm=TRUE)
sd(pap_all$confidence, na.rm=TRUE)
```

# Impute missing data

In the bacterial data, there are Acidobacteria and Chloroflexi taxa missing in the "...edge_data.csv" file. PAPRICA will not generate estimates if the taxa are not closely related enough to the representative taxa in it's database. To remedy this, I'm going to average the values for all other members in the same phylum and impute.

```{r}
# Missing data
pap_all %>% 
  filter(Phylum == "Acidobacteriota" | Phylum == "Chloroflexi")

# Average across phyla
missing_impute <- pap_all %>% 
  group_by(Phylum) %>% 
  summarize(n16s = mean(n16S, na.rm=TRUE),
            genome_size = mean(genome_size, na.rm=TRUE),
            phi = mean(phi, na.rm=TRUE),
            npaths_actual = mean(npaths_actual, na.rm=TRUE),
            nec_actual = mean(nec_actual, na.rm=TRUE)) %>% 
  filter(Phylum=="Acidobacteriota" | Phylum=="Chloroflexi")
missing_impute

# Acidobacteria imputations
acido_n16S <- as.numeric(missing_impute[1,2])
acido_genome <- as.numeric(missing_impute[1,3])
acido_phi <- as.numeric(missing_impute[1,4])
acido_npaths <- as.numeric(missing_impute[1,5])
acido_nec <- as.numeric(missing_impute[1,6])

# Chloroflexi imputations
chloro_n16S <- as.numeric(missing_impute[2,2])
chloro_genome <- as.numeric(missing_impute[2,3])
chloro_phi <- as.numeric(missing_impute[2,4])
chloro_npaths <- as.numeric(missing_impute[2,5])
chloro_nec <- as.numeric(missing_impute[2,6])


# Impute
pap_impute <- pap_all %>% 
  mutate(n16S = if_else(Phylum=="Acidobacteriota" & is.na(n16S), acido_n16S, if_else(Phylum=="Chloroflexi" & is.na(n16S), chloro_n16S, n16S)),
         genome_size = if_else(Phylum=="Acidobacteriota" & is.na(genome_size), acido_genome, 
                               if_else(Phylum=="Chloroflexi" & is.na(genome_size), chloro_genome, genome_size)),
         phi = if_else(Phylum=="Acidobacteriota" & is.na(phi), acido_phi, if_else(Phylum=="Chloroflexi" & is.na(phi), chloro_phi, phi)),
         npaths_actual = if_else(Phylum=="Acidobacteriota" & is.na(npaths_actual), acido_npaths, 
                                 if_else(Phylum=="Chloroflexi" & is.na(npaths_actual), chloro_npaths, npaths_actual)),
         nec_actual = if_else(Phylum=="Acidobacteriota" & is.na(nec_actual), acido_nec, 
                              if_else(Phylum=="Chloroflexi" & is.na(nec_actual), chloro_nec, nec_actual)))
```

# Output

Combine estimates with paprica predictions:

```{r}
growth_pap <- left_join(growth, pap_impute, by="ASV") %>% 
  select(-global_edge_num)

death_pap <- left_join(death, pap_impute, by="ASV") %>% 
  select(-global_edge_num)
```

Correct abundances using n16S prediction:

```{r}
growth_pap <- growth_pap %>%
  mutate(start_abund_corr = start_abund/n16S,
         end_abund_corr = end_abund/n16S,
         change_abund_corr = change_abund/n16S)

death_pap <- death_pap %>%
  mutate(start_abund_corr = start_abund/n16S,
         end_abund_corr = end_abund/n16S,
         change_abund_corr = change_abund/n16S)
```

Save:

```{r, eval=FALSE}
saveRDS(pap_impute, "../data_intermediate/SFA2_paprica_all2.rds")
saveRDS(growth_pap, "../data_intermediate/SFA2_growth_estimates_pap2.rds")
saveRDS(death_pap, "../data_intermediate/SFA2_death_estimates_pap2.rds")
```
