==> scripts/06_SFA2_CO2_processing.Rmd <==
---
title: "SFA2 - CO2 data processing"
author: "Cassandra Wattenburger"
date: "12/7/2020"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(warning = FALSE)
```

**Goal:** Analyze headspace CO2 from microcosms.

# Import libraries, clear working directory

```{r}
library(tidyverse)
library(cowplot)

sessionInfo()

rm(list=ls())
```

# Process GCMS data

## Import and reformat GCMS data

Note: Exported from Shimadzu PostRun Analysis software as ASCII using second and fourth output options.

This script is built to run on data that has been stored in a project directory containing a "data" directory with sub-directories for each time point that .txt files with the GCMS output for each sample.

```{r}
# List of directories containing data files
dir = list.dirs(path="../data_CO2/SFA2/data", full.names=TRUE, recursive=FALSE)

# Create lists of .txt files from directories
files = list()
for (i in dir) {
  files[[i]] <- list.files(path=i, pattern="*.txt", full.names=TRUE, recursive=FALSE)
}

# Extract area data
# TIC, w/z44 (12C), w/z45 (13C)

# Construct data frame containing all data
area_df = tibble() # initialize empty tibble

for (i in names(files)) { # for each list of files (refers to a directory)
  for (x in files[[i]]) { # for each file in the selected list (directory)
    # Create sample names based on file names
    # Strips the rest of the path from the sample name at the end
    name = str_remove(x, "../data_CO2/SFA2/data/.+/") %>%
      str_remove(".txt")
    #name.list[[i]] = append(name.list[[i]], name)
    
    # Isolate time point info
    date = str_remove(x, "../data_CO2/SFA2/data/") %>%
      str_remove("/.+.txt") %>%
      str_remove(".+_")
    tp = str_remove(x, "../data_CO2/SFA2/data/") %>%
      str_remove("/.+.txt") %>%
      str_remove("_.+")
    
    # Isolate the area data and clean up
    dat = read_tsv(x, skip=7, col_names=TRUE) %>%
      select("Measurement" = Name, Area) %>% # keep only the name of sample and the area measurements collected
      drop_na() %>%
      mutate(Measurement = case_when(Measurement == "TIC" ~ "Ctotal", # more explicit measurement labels
                              Measurement == "m/z 44" ~ "C12",
                              Measurement == "m/z 45" ~ "C13"),
             TP = tp, # adding additional metadata...
             Date = date,
             Sample = name)
    area_df = rbind(area_df, dat) # add to master data frame
  }
}

# Parse dates
area_df = area_df %>%
  mutate(Date = gsub("([0-2]{2})([0-9]{2})([0-2]{2})", "\\1-\\2-20\\3", Date)) %>%
  mutate(Date = parse_date(Date, "%m-%d-%Y"))
```

## Calculate total area

TIC area often does not record (not sure why), so we will use 12C + 13C instead which is a close approximation.

```{r}
# Calculate
area_df = area_df %>%
  pivot_wider(names_from = Measurement, values_from = Area) %>%
  mutate(area_C1213 = C12+C13)
```

## Standards

Note: some of these analyses were done while the experiment was ongoing, to adjust the number of standards run during the experiment.

Separate standards from samples/atmosphere measurements:

```{r}
std_df = area_df %>%
  filter(str_detect(Sample, "std"))
```

Correct split-ratios and add ppm of each standard

* standards 7-9 were run at a split-ratio of 90% rather than 50% because of the high CO2 concentration
* ppm calculated based on volume and % CO2 used and bottle volume

Multiplier: 0.5x = 0.9, x = 1.8

```{r}
# Calculate adjusted areas
std_df = std_df %>%
  mutate(area_adj = if_else(str_detect(Sample, "[7-9]"), area_C1213*1.8, area_C1213, missing=NULL), # correct split-ratio
         ppm = case_when(Sample == "std0" ~ 0, # add ppm
                         Sample == "std1" ~ 769,
                         Sample == "std2" ~ 1538,
                         Sample == "std3" ~ 3077,
                         Sample == "std4" ~ 7692,
                         Sample == "std5" ~ 15385,
                         Sample == "std6" ~ 30769,
                         Sample == "std7" ~ 76923,
                         Sample == "std7.5" ~ 153856,
                         Sample == "std8" ~ 230769,
                         Sample == "std8.5" ~ 307692,
                         Sample == "std9" ~ 384615)) 
  
```

Visualize:

```{r}
std_df %>%
  ggplot(aes(x=ppm, y=area_adj)) +
  geom_point() +
  geom_smooth(method="lm") +
  facet_wrap(~Date, scales="free") +
  theme_test()
```

Calculate calibration curve

```{r}
# Linear regressions
std_cal <- tibble()
for (x in unique(std_df$Date)) {
  dat <- filter(std_df, Date==x)
  std_lm <- lm(area_adj ~ ppm, dat)
  this_row <- bind_cols(Date = unique(dat$Date), # save results
                       intercept = summary(std_lm)$coefficients[1,1],
                       slope = summary(std_lm)$coefficients[2,1], 
                       adjR2 = summary(std_lm)$adj.r.squared, 
                       pval = summary(std_lm)$coefficients[2,4])
  std_cal <- bind_rows(std_cal, this_row)
}

std_cal
```

Remove std 1 - I think it would be better to include another high ppm sample instead of lower ones.

```{r}
std_df %>%
  filter(Sample != "std1") %>%
  ggplot(aes(x=ppm, y=area_adj)) +
  geom_point() +
  geom_smooth(method="lm") +
  facet_wrap(~Date, scales="free") +
  theme_test()
```

Calibration curve - no std1:

```{r}
# Linear regressions
std_cal1 <- tibble()
for (x in unique(std_df$Date)) {
  dat1 <- filter(std_df, Date==x & Sample != "std1")
  std_lm1 <- lm(area_adj ~ ppm, dat)
  date1 <- unique(dat1$Date)
  this_row1 <- bind_cols(Date = date, coeff = summary(std_lm1)$coefficients[2,1], adjR2 = summary(std_lm1)$adj.r.squared , pval = summary(std_lm1)$coefficients[2,4]) # save coefficent, adj. R2, and p-value
  std_cal1 <- bind_rows(std_cal, this_row)
}

std_cal1
```

Remove std 1 and 4:

```{r}
std_df %>%
  filter(!(Sample %in% c("std1", "std4"))) %>%
  ggplot(aes(x=ppm, y=area_adj)) +
  geom_point() +
  geom_smooth(method="lm") +
  facet_wrap(~Date, scales="free") +
  theme_test()
```

Calibration curve - no std1 or std4

```{r}
# Linear regressions
std_cal14 <- tibble()
for (x in unique(std_df$Date)) {
  dat14 <- filter(std_df, Date==x & Sample != "std1")
  std_lm14 <- lm(area_adj ~ ppm, dat14)
  this_row14 <- bind_cols(coeff = summary(std_lm14)$coefficients[2,1], adjR2 = summary(std_lm14)$adj.r.squared , pval = summary(std_lm14)$coefficients[2,4]) # save coefficent, adj. R2, and p-value
  std_cal14 <- bind_rows(std_cal14, this_row14)
}

std_cal14
```

Same result. Will add more high ppm standards and omit stds 1 and 4 for future runs.

## Samples

Isolate microcosm samples:

```{r}
sample_df <- anti_join(area_df, std_df) %>%
  mutate(dummy_ppm = 0) %>% # dummy ppm for graphing with standards on ppm axis
  filter(!str_detect(Sample, "[a-z]")) %>% # exclude other non-samples 
  mutate(Sample, Sample = parse_number(Sample)) # convert to numeric
```

Add metadata:

```{r}
# Import metadata
meta_sample <- read_csv("../data_CO2/SFA2/SFA2_headspace_metadata.csv") %>% 
  mutate(Inoculant = as.character(Inoculant)) %>% 
  select(Sample, Inoculant)

meta_tp <- read_tsv("../data_CO2/SFA2/headspace_tps_metadata.tsv")

# Merge metadata with CO2 data
sample_df <- full_join(sample_df, meta_sample, by="Sample") %>% 
  mutate(TP = gsub("tp", "", TP),
         TP = as.numeric(TP)) %>% 
  left_join(meta_tp, by="TP") %>% 
  select(Sample, Inoculant, TP, Date, Day, Timeframe, area_C1213, dummy_ppm)
```

Visualize:

```{r}
# Compare samples against range of standards
std_df %>%
  ggplot(aes(x=ppm, y=area_adj)) +
  geom_point(shape=1) +
  geom_point(aes(y=area_C1213, x=dummy_ppm, color=Date), data=sample_df) +
  labs(title="Samples vs. standards") +
  theme_test()

# Observe progression of CO2 production
sample_df %>%
  mutate(Control = if_else(Sample %in% c(379:384), "Y", "N")) %>% # distinguish negative control
  ggplot(aes(y=area_C1213, x=Date)) +
  labs(y="Area") +
  geom_jitter(aes(color = Control)) +
  labs(title="Quick look - respiration over time") +
  theme_test()
```

Outlier ~ Dec. 15, looking at lab notes it seems I may have accidentally sampled twice into the same vial for this sample. This means this value is ~2x more than what it should be. I'll correct this by halving the area value to approximate the correct measurement.

```{r}
# Isolate outlier and correct
outlier <- sample_df %>%
  subset(Sample==363 & TP==6) %>%
  mutate(area_C1213 = area_C1213/2) # halve measurement

# Replace in dataframe
sample_df <- sample_df %>%
  filter(!(Sample==363 & TP==6)) %>%
  bind_rows(outlier)
```

## Calculate CO2

Use calibration curve to convert sample areas into ppm.

```{r}
# Convert sample area to ppm using calibration curves
sample_conv <- tibble()
for (d in unique(sample_df$Date)) {
  cal = std_cal %>% filter(Date==d)
  conv = sample_df %>% 
    filter(Date==d) %>%
    mutate(ppm = (area_C1213 - cal$intercept)/cal$slope)
  sample_conv = rbind(sample_conv, conv)
}
```

Sanity check visualization: Do the sample measurements fall within the standards?

```{r}
# Sanity check, plot against standards
std_df %>%
  ggplot(aes(x=ppm, y=area_adj)) +
  geom_point(shape=1) +
  geom_point(aes(y=area_C1213, x=ppm, color=Date), data=sample_conv) +
  #scale_y_continuous(limits=c(0,500000)) +
  #scale_x_continuous(limits=c(0, 20000)) +
  theme_test() +
  theme()
```

Calculate mg CO2:

```{r}
# Ideal gas law: n=(P*V)/(R*temp)
# NOTE: pressure does not matter for this calulation because it's all relative to the stds
R <- 0.082057338 # universal gas constant, (L*atm)/(mol*K)
temp <- 296.48 # K, 74F in lab throughout experiment
V <- 0.013 # L, headspace of 30 mL serum bottle with sand
mwCO2 <- 44.01 # molecular weight (g/mol) of CO2
P <- 1 #atm, assumes 1 atm in ucosms during sampling

# Calculate per time point
molCO2100 <- P*V/(R*temp) # total mol of CO2 in headspace, if it were 100% CO2
mgCO2100 <- molCO2100 * mwCO2*1000 # mol * g/mol * 1000mg/g = mg, mg of CO2 present if it were 100% CO2
sample_conv <- mutate(sample_conv, mgCO2 = mgCO2100*(ppm*10^-6)) # actual mg CO2 based on ppm
```

## Save processed CO2 data for use in other scripts

```{r, eval=FALSE}
resp_df <- sample_conv %>% 
  select(Sample, Inoculant, Day, Timeframe, ppm, mgCO2)

saveRDS(resp_df, file="../data_intermediate/SFA2_respiration.rds")
```
